requirements.txt
pygame
pytest
neat-python